Create new Rails app
The app was created with this command:

    rails _7.1.5.1_ new quote-editor --css=sass --javascript=esbuild --database=postgresql

I'm following the Turbo Rails tutorial [1] which was written for Rails
7.0, and that recommends using turbo-rails 1.x (the current version at
the time 7.0 was released).

I would have created this app with 7.0.8.7 (the latest 7.0 release as I
write), but it threw an exception while creating a new app that I didn't
fancy investigating.

[1] https://www.hotrails.dev/turbo-rails/turbo-rails-tutorial-introduction