Delete quotes via a Turbo Stream
The key bit of magic that you can't see in this code, is that Rails 7+
submits forms on pages with Turbo enabled using the TURBO_STREAM format.

This means that we can respond a snippet of markup (a <turbo-stream>
tag) that will instruct the Turbo JavaScript library to take some action
on a turbo frame (in this case, we're removing it).

As covered here:

https://www.hotrails.dev/turbo-rails/turbo-frames-and-turbo-streams